moved logging into logging module
