+++ src/logging.js
@@ -8,43 +8,82 @@
  *  @license    Licensed under the GPLv3 license.
  **/
 
-import { createLogger, format, transports } from 'winston'
+import { createLogger, format, transports } from "winston";
 
-const logFormat = format.combine(
-  format.timestamp(),
-  format.printf(({level, message, timestamp}) => (
-      `${timestamp} ${level}: ${message}`
-  ))
-);
+import { getEnvAsNumber } from "./utils";
+
+const LOG_LEVEL = getEnvAsNumber("LOG_LEVEL");
 
 var logger;
 
-if (process.env.NODE_ENV === 'development') {
+const logFormat = format.combine(
+  format.timestamp(),
+  format.printf(
+    ({ level, message, timestamp }) => `${timestamp} ${level}: ${message}`
+  )
+);
+
+if (process.env.NODE_ENV === "development" || LOG_LEVEL === 1) {
   logger = createLogger({
-    level: 'debug',
+    level: "debug",
     format: logFormat,
     transports: [
       new transports.File({
-        level: 'debug',
-        filename: 'combined.log'
+        level: "debug",
+        filename: "combined.log"
       })
     ]
   });
   //logger.add(new transports.Console({
-    //colorize: 'all',
-    //format: format.simple()
+  //colorize: 'all',
+  //format: format.simple()
   //}));
 } else {
   logger = createLogger({
-    level: 'error',
+    level: "error",
     format: logFormat,
     transports: [
       new transports.File({
-        level: 'error',
-        filename: 'error.log'
+        level: "error",
+        filename: "error.log"
       })
     ]
   });
 }
+
+/**
+ *  logging of state-store messages
+ **/
+const stateTransformer = function(state) {
+  const toPrint = {};
+  toPrint.sequencers = state.sequencers;
+  //Object.keys(state.sequencers).forEach(function (seqId) {
+  //toPrint.sequencers[seqId] = _.pick(state.sequencers[seqId], [
+  //'type',
+  //'playingState',
+  //'playQuant',
+  //'stopQuant',
+  //'event',
+  //'bufSequence'
+  //])
+  //});
+  toPrint.sessionPhase = state.sessionPhase;
+  //toPrint.sessionPhaseDurations = state.sessionPhaseDurations;
+  toPrint.idlePlayer = state.idlePlayer;
+  toPrint.soundReady = state.soundReady;
+  toPrint.songId = state.songId;
+  //toPrint.sequencers = state.sequencers;
+  //toPrint.segments = state.segments;
+  return toPrint;
+};
+
+export const storeLogger = store => next => action => {
+  logger.info("action: " + JSON.stringify(action));
+  const result = next(action);
+  logger.info(
+    "next state: " + JSON.stringify(stateTransformer(store.getState()), " ", 4)
+  );
+  return result;
+};
 
 export default logger;
